Move code.go and code_test.go inside internal module
